Completed reformatting; now espresso calculators are all grouped together in a folder
