+++ ase/io/espresso/__init__.py
@@ -4,8 +4,9 @@ Written by Edward Linscott 2020-21
 
 """
 
-from _pw import read_espresso_in, read_espresso_out, write_espresso_in
-from _koopmans_cp import read_koopmans_cp_in, read_koopmans_cp_out
-from _koopmans_screen import read_koopmans_screen_in, read_koopmans_screen_out
-from _koopmans_ham import read_koopmans_ham_in, read_koopmans_ham_out
-from _pw2wannier import read_pw2wannier_in, read_pw2wannier_out
+from .pw import read_espresso_in, read_espresso_out, write_espresso_in
+from .koopmans_cp import read_koopmans_cp_in, read_koopmans_cp_out, write_koopmans_cp_in
+from .koopmans_screen import read_koopmans_screen_in, read_koopmans_screen_out, write_koopmans_screen_in
+from .koopmans_ham import read_koopmans_ham_in, read_koopmans_ham_out, write_koopmans_ham_in
+from .pw2wannier import read_pw2wannier_in, read_pw2wannier_out, write_pw2wannier_in
+from .wann2kc import read_wann2kc_in, read_wann2kc_out, write_wann2kc_in
