+++ ase/io/espresso/pw2wannier.py
@@ -4,8 +4,8 @@
 
 from ase.utils import basestring
 from ase.atoms import Atoms
-from ase.calculators.pw2wannier import PW2Wannier
-from ase.io.espresso._utils import Namelist, read_fortran_namelist
+from ase.calculators.espresso import PW2Wannier
+from .utils import Namelist, read_fortran_namelist
 
 
 def read_pw2wannier_in(fileobj):
